Extract inner modules of `gpio` module into their own files
