+++ examples/src/bin/lp_core_basic.rs
@@ -12,7 +12,7 @@
 
 use esp_backtrace as _;
 use esp_hal::{
-    gpio::{lp_gpio::IntoLowPowerPin, IO},
+    gpio::{lp_io::IntoLowPowerPin, IO},
     lp_core::{LpCore, LpCoreWakeupSource},
     peripherals::Peripherals,
     prelude::*,
