+++ examples/src/bin/lp_core_uart.rs
@@ -13,7 +13,7 @@
 use esp_backtrace as _;
 use esp_hal::{
     clock::ClockControl,
-    gpio::{lp_gpio::IntoLowPowerPin, IO},
+    gpio::{lp_io::IntoLowPowerPin, IO},
     lp_core::{LpCore, LpCoreWakeupSource},
     peripherals::Peripherals,
     prelude::*,
